+++ bin/node/src/commands/bundled.rs
@@ -10,9 +10,9 @@ use tokio::{net::TcpListener, task::JoinSet};
 use url::Url;
 
 use super::{
-    DEFAULT_BATCH_INTERVAL_MS, DEFAULT_BLOCK_INTERVAL_MS, DEFAULT_MONITOR_INTERVAL_MS,
-    DEFAULT_NTX_TICKER_INTERVAL_MS, ENV_BATCH_PROVER_URL, ENV_BLOCK_PROVER_URL, ENV_DATA_DIRECTORY,
-    ENV_ENABLE_OTEL, ENV_NTX_PROVER_URL, ENV_RPC_URL, parse_duration_ms,
+    DEFAULT_BATCH_INTERVAL, DEFAULT_BLOCK_INTERVAL, DEFAULT_MONITOR_INTERVAL,
+    DEFAULT_NTX_TICKER_INTERVAL, ENV_BATCH_PROVER_URL, ENV_BLOCK_PROVER_URL, ENV_DATA_DIRECTORY,
+    ENV_ENABLE_OTEL, ENV_NTX_PROVER_URL, ENV_RPC_URL, duration_to_human_readable_string,
 };
 use crate::system_monitor::SystemMonitor;
 
@@ -73,39 +73,39 @@ pub enum BundledCommand {
         #[arg(long = "enable-otel", default_value_t = false, env = ENV_ENABLE_OTEL, value_name = "BOOL")]
         open_telemetry: bool,
 
-        /// Interval at which to produce blocks in milliseconds.
+        /// Interval at which to produce blocks.
         #[arg(
             long = "block.interval",
-            default_value = DEFAULT_BLOCK_INTERVAL_MS,
-            value_parser = parse_duration_ms,
-            value_name = "MILLISECONDS"
+            default_value = &duration_to_human_readable_string(DEFAULT_BLOCK_INTERVAL),
+            value_parser = humantime::parse_duration,
+            value_name = "DURATION"
         )]
         block_interval: Duration,
 
         /// Interval at which to run the network transaction builder's ticker.
         #[arg(
             long = "ntb.interval",
-            default_value = DEFAULT_NTX_TICKER_INTERVAL_MS,
-            value_parser = parse_duration_ms,
-            value_name = "MILLISECONDS"
+            default_value = &duration_to_human_readable_string(DEFAULT_NTX_TICKER_INTERVAL),
+            value_parser = humantime::parse_duration,
+            value_name = "DURATION"
         )]
         ntx_ticker_interval: Duration,
 
-        /// Interval at which to procude batches in milliseconds.
+        /// Interval at which to produce batches.
         #[arg(
             long = "batch.interval",
-            default_value = DEFAULT_BATCH_INTERVAL_MS,
-            value_parser = parse_duration_ms,
-            value_name = "MILLISECONDS"
+            default_value = &duration_to_human_readable_string(DEFAULT_BATCH_INTERVAL),
+            value_parser = humantime::parse_duration,
+            value_name = "DURATION"
         )]
         batch_interval: Duration,
 
-        /// Interval at which to monitor the system in milliseconds.
+        /// Interval at which to monitor the system.
         #[arg(
             long = "monitor.interval",
-            default_value = DEFAULT_MONITOR_INTERVAL_MS,
-            value_parser = parse_duration_ms,
-            value_name = "MILLISECONDS"
+            default_value = &duration_to_human_readable_string(DEFAULT_MONITOR_INTERVAL),
+            value_parser = humantime::parse_duration,
+            value_name = "DURATION"
         )]
         monitor_interval: Duration,
     },